+++ react/ComparisonGrid.tsx
@@ -1,0 +1,3 @@
+import ComparisonGrid from './components/comparisonList/ComparisonGrid'
+
+export default ComparisonGrid